+++ CHANGELOG.md
@@ -40,15 +40,9 @@
   keyword when it can be omitted inside closures.  
   [Marcelo Fabri](https://github.com/marcelofabri)
   [#1194](https://github.com/realm/SwiftLint/issues/1194)
-* Add opt-in option to `unused_optional_binding` rule to ignore `try?` in `guard` statements  
-  [Sega-Zero](https://github.com/Sega-Zero)
-  [#1432](https://github.com/realm/SwiftLint/issues/1432)
   
-  
-* Add lowercase and missing colon checks to the `mark` rule.  
-  [Jason Moore](https://github.com/xinsight)
-  
-* Add opt-in option to `unused_optional_binding` rule to ignore `try?` in `guard` statements  
+* Add option to `unused_optional_binding` rule to ignore `try?`
+  in `guard` statements.  
   [Sega-Zero](https://github.com/Sega-Zero)
   [#1432](https://github.com/realm/SwiftLint/issues/1432)
